[#7] clap 구성
