+++ Cargo.toml
@@ -17,3 +17,4 @@ homepage = "https://github.com/myyrakle/clockpipe/blob/master/README.md"
 async-trait = "0.1.88"
 tokio = { version="1.46.1", features=["full"] }
 clap = { version = "4.4.18", features = ["derive"] }
+serde = { version = "1", features = ["derive"] }
